added: alert for error
